feat(master): fix linter format

+++ contracts/AbiDecode.sol
@@ -2,30 +2,33 @@
 pragma solidity ^0.8.16;
 
 contract AbiDecode {
-    struct MyStruct {
-        string name;
-        uint[2] nums;
-    }
+  struct MyStruct {
+    string name;
+    uint256[2] nums;
+  }
 
-    function encode(
-        uint x,
-        address addr,
-        uint[] calldata arr,
-        MyStruct calldata myStruct
-    ) external pure returns (bytes memory) {
-        return abi.encode(x, addr, arr, myStruct);
-    }
+  function encode(
+    uint256 x,
+    address addr,
+    uint256[] calldata arr,
+    MyStruct calldata myStruct
+  ) external pure returns (bytes memory) {
+    return abi.encode(x, addr, arr, myStruct);
+  }
 
-    function decode(bytes calldata data)
-        external
-        pure
-        returns (
-            uint x,
-            address addr,
-            uint[] memory arr,
-            MyStruct memory myStruct
-        )
-    {
-        (x, addr, arr, myStruct) = abi.decode(data, (uint, address, uint[], MyStruct));
-    }
+  function decode(bytes calldata data)
+    external
+    pure
+    returns (
+      uint256 x,
+      address addr,
+      uint256[] memory arr,
+      MyStruct memory myStruct
+    )
+  {
+    (x, addr, arr, myStruct) = abi.decode(
+      data,
+      (uint256, address, uint256[], MyStruct)
+    );
+  }
 }
